refactor: extract PostCard component
