+++ pages/category/[category].js
@@ -1,34 +1,32 @@
 /* eslint jsx-a11y/anchor-is-valid: 0 */
-import React, { useState } from "react"
-import Link from "next/link"
+import React, { useState } from 'react'
+import Link from 'next/link'
 import {
   Container,
   Row,
   Col,
-  Card,
-  CardBody,
   Button,
   Badge,
-  CardFooter,
-  Tooltip,
   Form,
   FormGroup,
   FormInput,
   Alert,
-} from "shards-react"
-import PageTitle from "../../components/common/PageTitle"
-import PageMetadata from "../../components/common/Helmet"
-import http from "../../services/Apicalls"
+} from 'shards-react'
+import PageTitle from '../../components/common/PageTitle'
+import PageMetadata from '../../components/common/Helmet'
+import http from '../../services/Apicalls'
+import PostCard from '../../components/PostCard'
+import { trimmedPostBody } from '../../utils/trimmedPostBody'
 
 function CategoryPosts({ posts, category, allPosts }) {
   const [reader, setReader] = useState({
-    email: "",
+    email: '',
   })
   const [subscribedState, setSubscribedState] = useState({
     isSubmit: false,
     alert: false,
-    alertMessage: "",
-    alertColor: "",
+    alertMessage: '',
+    alertColor: '',
   })
 
   const postIdList = posts.map((post) => ({
@@ -36,10 +34,10 @@ function CategoryPosts({ posts, category, allPosts }) {
     tooltipOpen: false,
   }))
 
-  const postsArray = [...posts]
-  const categories = postsArray
-    .map((item) => item.category)
-    .filter((v, i, self) => i == self.indexOf(v))
+  // const postsArray = [...posts]
+  // const categories = postsArray
+  //   .map((item) => item.category)
+  //   .filter((v, i, self) => i == self.indexOf(v))
   const tags = allPosts
     .map((item) => item.tags)
     .flat()
@@ -47,16 +45,6 @@ function CategoryPosts({ posts, category, allPosts }) {
   const popularPosts = allPosts
     .sort((a, b) => b.comments.length - a.comments.length)
     .slice(0, 3)
-
-  const [postToggleOpen, setPostToggleOpen] = useState(postIdList)
-  const toggleUniqueTooltip = (tooltipId) => {
-    const data = [...postToggleOpen]
-    data[tooltipId] = {
-      ...data[tooltipId],
-      tooltipOpen: !data[tooltipId].tooltipOpen,
-    }
-    setPostToggleOpen(data)
-  }
 
   const onChange = (e) => {
     setReader({
@@ -71,18 +59,18 @@ function CategoryPosts({ posts, category, allPosts }) {
     })
     e.preventDefault()
     http
-      .post("/api/subscribers", reader)
+      .post('/api/subscribers', reader)
       .then((res) => {
         if (res.status === 201) {
           setSubscribedState({
             ...subscribedState,
             alert: true,
-            alertColor: "success",
+            alertColor: 'success',
             alertMessage: res.data.message,
           })
-          setReader({ email: "" })
+          setReader({ email: '' })
         } else {
-          let errorText = ""
+          let errorText = ''
           if (res.data.errors)
             res.data.errors.forEach((error) => {
               errorText += error.msg
@@ -91,7 +79,7 @@ function CategoryPosts({ posts, category, allPosts }) {
           setSubscribedState({
             ...subscribedState,
             alert: true,
-            alertColor: "warning",
+            alertColor: 'warning',
             alertMessage: errorMessage,
           })
         }
@@ -102,18 +90,6 @@ function CategoryPosts({ posts, category, allPosts }) {
           setSubscribedState({ alert: false })
         }, 5000)
       })
-  }
-
-  const trimmedPostBody = (postBody, maxChar) => {
-    // const maxChar = 150
-    if (postBody.length > maxChar) {
-      postBody = postBody.substring(0, maxChar)
-      postBody = `${postBody.substring(
-        0,
-        Math.min(postBody.length, postBody.lastIndexOf(" "))
-      )} . . .`
-    }
-    return postBody
   }
 
   return (
@@ -131,52 +107,12 @@ function CategoryPosts({ posts, category, allPosts }) {
         <Col md="9" lg="9">
           <Row>
             {posts.map((post, idx) => (
-              <Col lg="6" md="6" sm="12" className="mb-4" key={idx}>
-                <Card small className="card-post card-post--1">
-                  <div
-                    className="card-post__image"
-                    style={{ backgroundImage: `url(${post.postImage})` }}
-                  >
-                    <Badge pill className="card-post__category bg-primary">
-                      {post.category}
-                    </Badge>
-                  </div>
-                  <CardBody id={`post-${post._id}`}>
-                    <h4 className="card-title">
-                      <Link
-                        href={`/post/${post.slug}`}
-                        as={`/post/${post.slug}`}
-                      >
-                        <a className="text-fiord-blue">
-                          {trimmedPostBody(post.title, 50)}
-                        </a>
-                      </Link>
-                    </h4>
-                    <div
-                      dangerouslySetInnerHTML={{
-                        __html: trimmedPostBody(post.body, 120),
-                      }}
-                    />
-                  </CardBody>
-                  <Tooltip
-                    style={{
-                      minWidth: "320px",
-                      fontSize: "20px",
-                    }}
-                    // placement="bottom"
-                    open={postToggleOpen[idx].tooltipOpen}
-                    target={`#post-${post._id}`}
-                    toggle={() => toggleUniqueTooltip(idx)}
-                  >
-                    {post.title}
-                  </Tooltip>
-                  <CardFooter>
-                    <span className="text-muted">
-                      {new Date(post.updatedAt).toDateString()}
-                    </span>
-                  </CardFooter>
-                </Card>
-              </Col>
+              <PostCard
+                post={post}
+                key={post._id}
+                postIdList={postIdList}
+                idx={idx}
+              />
             ))}
           </Row>
         </Col>
@@ -237,9 +173,9 @@ function CategoryPosts({ posts, category, allPosts }) {
               <img
                 src={post.postImage}
                 style={{
-                  objectFit: "cover",
-                  maxWidth: "100%",
-                  height: "100px",
+                  objectFit: 'cover',
+                  maxWidth: '100%',
+                  height: '100px',
                 }}
                 alt="post"
               />
@@ -292,32 +228,12 @@ function CategoryPosts({ posts, category, allPosts }) {
   )
 }
 
-export async function getStaticPaths() {
+export async function getServerSideProps({ params: { category } }) {
   // Call an external API endpoint to get posts
-  const res = await http.get(`/api/post`)
-  const posts = await res.data.posts
-
-  // Get the paths we want to pre-render based on posts
-  // const paths = posts.map((post) => `/post/${post._id}`)
-  const categories = posts
-    .map((item) => item.category)
-    .filter((v, i, self) => i == self.indexOf(v))
-
-  const paths = categories.map((category) => ({
-    params: { category },
-  }))
-  // We'll pre-render only these paths at build time.
-  // { fallback: false } means other routes should 404.
-  return { paths, fallback: false }
-}
-
-export async function getStaticProps({ params }) {
-  // Call an external API endpoint to get posts
-  const res = await http.get(`/api/posts/categories/${params.category}`)
-  const postsData = await http.get("/api/post")
+  const res = await http.get(`/api/posts/categories/${category}`)
+  const postsData = await http.get('/api/post')
   const posts = await res.data.posts.reverse()
   const allPosts = await postsData.data.posts.reverse()
-  const { category } = params
 
   // By returning { props: posts }, the Blog component
   // will receive `post` as a prop at build time
